docmind/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
httpx==0.27.2

# Document Processing
pymupdf==1.24.11
easyocr==1.7.2
Pillow==10.4.0

# Embeddings
sentence-transformers==3.2.1
torch>=2.1.0
transformers>=4.44.0

# Vector Store
faiss-cpu==1.8.0

# Frontend
streamlit==1.39.0
streamlit-chat==0.1.1

# Utilities
numpy==1.26.4
python-dotenv==1.0.1
tqdm==4.67.0
